+++ changelog.md
@@ -5,6 +5,10 @@ This change log records updates after 2025-3-24.
 
 Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more info.
 
+## [v0.5.1] - 2025-4-18
+### Changed
+- Modularize `layer` into separate `dense` and `simple_rnn` modules
+
 ## [v0.5.1] - 2025-4-17
 ### Added
 - Add `SimpleRNN` layer initial implementation